Fix sync from wechatwork
