+++ pkg/cmd/wechatwork.go
@@ -21,18 +21,9 @@
 package cmd
 
 import (
-	"fmt"
-	"net/url"
-	"strings"
-	"time"
-
 	"github.com/spf13/cobra"
 
-	"github.com/wealthworks/go-tencent-api/exwechat"
-
-	"github.com/liut/staffio/pkg/backends"
-	"github.com/liut/staffio/pkg/models"
-	"github.com/liut/staffio/pkg/models/team"
+	"github.com/liut/staffio/pkg/backends/wechatwork"
 )
 
 // wechatworkCmd represents the wechatwork command
@@ -40,13 +31,6 @@ var wechatworkCmd = &cobra.Command{
 	Use:   "wechatwork",
 	Short: "Sync with wechat work",
 	Run:   wechatworkRun,
-}
-var nameReplacer = strings.NewReplacer("公司", "", "总部", "", "分公司", "", "项目组", "")
-
-type wStaff struct {
-	models.Staff
-	Leader bool `json:"leader,omitempty" form:"-"` // temporary var
-	TeamID int  `json:"teamID,omitempty" form:"-"` // department id
 }
 
 func init() {
@@ -69,118 +53,5 @@ func wechatworkRun(cmd *cobra.Command, args []string) {
 	action, _ := cmd.Flags().GetString("act")
 	uid, _ := cmd.Flags().GetString("uid")
 
-	log := logger()
-
-	wechat := exwechat.New(settings.WechatCorpID, settings.WechatContactSecret)
-	if action == "query" {
-		if len(uid) > 0 {
-			user, err := wechat.GetUser(uid)
-			if err != nil {
-				log.Warnw("wechat GetUser fail", "err", err)
-				return
-			}
-			fmt.Println(user)
-			return
-		}
-		fmt.Println("empty uid")
-		return
-	}
-
-	departments, err := wechat.ListDepartment(1)
-	if err != nil {
-		log.Infow("wechat ListDepartment fail", "err", err)
-		return
-	}
-	// log.Infof("departments: %v", data)
-	svc := backends.NewService()
-	for _, dept := range departments {
-		// fmt.Printf("%4d  %10s   %d   %d\n", dept.Id, dept.Name, dept.ParentId, dept.Order)
-		users, err := wechat.ListUser(dept.Id, false)
-		if err != nil {
-			log.Infow("wechat ListUser fail", "err", err)
-			return
-		}
-		name := dept.Name
-
-		if parent := departments.WithID(dept.ParentId); parent != nil {
-			name = nameReplacer.Replace(parent.Name) + "-" + dept.Name
-		}
-		var team = &team.Team{
-			ID:      dept.Id,
-			Name:    name,
-			Updated: time.Now(),
-		}
-		var staffs []*wStaff
-		for _, val := range users {
-			if !val.IsActived() || !val.IsEnabled() {
-				log.Infow("invalid user status", "user", val.Name, "status", val.Status, "enabled", val.Enabled)
-				continue
-			}
-			team.Members = append(team.Members, val.UID)
-			staff := userToStaff(&val)
-			if val.IsLeader == 1 {
-				team.Leaders = append(team.Leaders, staff.UID)
-			}
-			// fmt.Println(staff)
-			staffs = append(staffs, staff)
-			// fmt.Printf("%4s %10s  %v\n", val.UID, val.Name, val.DepartmentIds)
-		}
-
-		fmt.Printf("%4d  %12s   %v \n", team.ID, team.Name, team.Members)
-		if strings.HasPrefix(action, "sync") {
-			var leader string
-			if action == "sync-all" {
-				for _, staff := range staffs {
-					if staff.Leader {
-						err = svc.Team().AddManager(dept.Id, staff.UID)
-						if err != nil {
-							log.Infow("add manager fail", "uid", staff.UID, "id", dept.Id, "err", err)
-						}
-						leader = staff.UID
-					}
-					if err = svc.SaveStaff(&staff.Staff); err != nil {
-						fmt.Printf("save staff %v ERR %s\n", staff, err)
-						if err != backends.ErrEmptyEmail {
-							return
-						}
-					}
-					fmt.Printf("save staff %s(%s) %q OK\n", staff.CommonName, staff.UID, leader)
-				}
-			}
-			err = svc.Team().Store(team)
-			if err == nil {
-				fmt.Printf("saved department(%q, %q) to team OK\n", team.Name, team.Leaders)
-			} else {
-				fmt.Printf("save department %s ERR %s\n", dept.Name, err)
-				return
-			}
-		}
-
-	}
-}
-
-func userToStaff(user *exwechat.User) *wStaff {
-	staff := &wStaff{
-		Staff: models.Staff{
-			UID:          strings.ToLower(user.UID),
-			CommonName:   user.Name,
-			Email:        user.Email,
-			Mobile:       user.Mobile,
-			Gender:       models.Gender(user.Gender).String(),
-			EmployeeType: user.Title,
-		},
-		Leader: user.IsLeader == 1,
-	}
-	staff.Surname, staff.GivenName = models.SplitName(user.Name)
-	if user.Avatar != "" {
-		uri, err := url.Parse(user.Avatar)
-		if err == nil {
-			staff.AvatarPath = uri.Path
-		}
-	}
-	if user.Alias != "" {
-		staff.Nickname = user.Alias
-	}
-
-	return staff
+	wechatwork.SyncDepartment(action, uid)
 }
